Improved admin menu and settings page
The buttons on the admin menu are displayed related to the page, but the
system was based on the index of the buttons in the adminBtns array. So
I added a unique key used with a method called getAdminBtnByKey, so the
order can be changed without corrupting the menu behavior.

On the "settings" page, "pages" tab, I have integrated the Draggable
component to reorder pages with drag and drop events. Missing the update
in database.