Medium Editor
Switched custom style from custom tags to custom classes.

Created encoder vue plugin allowing to make an editor of content in a textarea. It works like BBCode or Wordpress editor with custom tags like "[b]bold text[/b]" It must be implemented by a component showing the toolbar and a textarea to be modified. Heavy solution which can not support contentEditable HTML5 attribute yet.
This plugin is deprecated for this website and has been replaced by Medium Editor.

Added Medium Editor and the associated wrapping vue package named Vue Medium Editor.
Created an extension for this editor called animate:

It allows to apply an animation on the selected text when in the editor. The modification of an already defined animation is to be made.

Separated the editor part from the page form so that component can be implemented everywhere